documentprocessingnew-main copy/docpro_agent/requirements.txt
requests
watchdog
pystray
Pillow
winshell
pywin32
pyinstaller
